+++ services/chaindb/postgresql/upgrader.go
@@ -108,6 +108,7 @@ var upgrades = map[uint64]*upgrade{
 	12: {
 		funcs: []func(context.Context, *Service) error{
 			addValidatorWithdrawalCredentials,
+			addValidatorIndexToChanges,
 		},
 	},
 }
@@ -1550,15 +1551,32 @@ func addValidatorWithdrawalCredentials(ctx context.Context, s *Service) error {
 	}
 
 	if _, err := tx.Exec(ctx, `
-	ALTER TABLE t_validators ADD COLUMN f_withdrawal_credentials BYTEA;
+ALTER TABLE t_validators
+ADD COLUMN f_withdrawal_credentials BYTEA
 `); err != nil {
-		return errors.Wrap(err, "failed to create block bls to execution changes table")
+		return errors.Wrap(err, "failed to add f_withdrawal_credentials to t_validators")
 	}
 
 	if _, err := tx.Exec(ctx, `
-CREATE INDEX IF NOT EXISTS i_validators_3 ON t_validators(f_withdrawal_credentials);
+CREATE INDEX IF NOT EXISTS i_validators_3 ON t_validators(f_withdrawal_credentials)
 `); err != nil {
-		return errors.Wrap(err, "failed to create validators index 3")
+		return errors.Wrap(err, "failed to create i_validators_3")
+	}
+
+	return nil
+}
+
+func addValidatorIndexToChanges(ctx context.Context, s *Service) error {
+	tx := s.tx(ctx)
+	if tx == nil {
+		return ErrNoTransaction
+	}
+
+	if _, err := tx.Exec(ctx, `
+ALTER TABLE t_block_bls_to_execution_changes
+ADD COLUMN f_validator_index BIGINT
+`); err != nil {
+		return errors.Wrap(err, "failed to add f_validator_index to t_block_bls_to_execution_changes")
 	}
 
 	return nil